# docs/openapi/api-internal-v1.yaml
openapi: 3.0.3
info:
  termsOfService: https://termofservice.it
  title: 'Stato della piattaforma PN'
  x-summary: 'Stato della piattaforma PN'
  version: '1.0.0'
  description: >- 
    Queste API permettono:<br/>
    - la visualizzazione dello stato dei servizi di Piattaforma Notifiche 
    e dello storico dei disservizi; per ogni disservizio è possibile scaricare il corrispondente
    atto opponibile a terzi corrispondente. <br/>
    - l'inserimento di eventi di inizio o fine di un disservizio utile in quei casi in cui il 
    sistema automatico non rivela il problema. <br/>
    __N.B.:__ gli atti opponibili a terzi vengono generati solo per i disservizi conclusi.
  contact:
    email: pn@pagopa.it
  license:
    name: Licenza di PN
    url: 'https://da-definire/'
servers:
- url: https://api.notifichedigitali.it
  description: Ambiente di produzione
- url: https://api.uat.notifichedigitali.it
  description: Ambiente di test
- url: https://api.dev.notifichedigitali.it
  description: Ambiente di sviluppo
tags:
  - name: HealthCheck                               # NO EXTERNAL
    description: >-                                 # NO EXTERNAL
      Invocazioni per sapere lo stato di PN         # NO EXTERNAL
  - name: Status
    description: >-
      Lettura dello stato di PN, API pubbliche
  - name: Downtime
    description: >-
      Invocazioni per recupero dei disservizi di PN
  - name: DowntimeInternal                                                  # NO EXTERNAL
    description: >-                                                         # NO EXTERNAL
      Aggiornamento dello stato di PN, API sottoposte ad autenticazione.    # NO EXTERNAL
  - name: interopProbing
    description: >-
      Invocazioni per sapere lo stato dell'E-service

paths:
    ##########################################################################################
    ###                           API DI STATO RICHIESTE DA AGID                           ###
    ##########################################################################################
  
  "/healthcheck":                                                                                       # NO EXTERNAL
    get:                                                                                                # NO EXTERNAL         
      summary: healthCheck path                                                                         # NO EXTERNAL      
      description: health check path per verificare lo stato del micro servizio                         # NO EXTERNAL
      tags:                                                                                             # NO EXTERNAL
        - HealthCheck                                                                                   # NO EXTERNAL
      operationId: healthcheck                                                                          # NO EXTERNAL
      responses:                                                                                        # NO EXTERNAL
        '200':                                                                                          # NO EXTERNAL
          description: Ok                                                                               # NO EXTERNAL
        '500':                                                                                          # NO EXTERNAL
          description: Internal Server Error                                                            # NO EXTERNAL
    ###########################################################################################
    ###                                 LETTURA DELLO STATO                                 ###
    ###########################################################################################
  "/status":                                                                                            # ONLY EXTERNAL
    get:                                                                                                # ONLY EXTERNAL
      summary: status path                                                                         # ONLY EXTERNAL
      description: >-                                                                                   
        Questo path e questo metodo devono, per richiesta di AGID, restituire <br/>                    
          - status 200 se la piattaforma notifica è correttamente funzionante, <br/>                    
          - status 500 se la piattaforma ha un funzionamento degradato. <br/>                          
        Il body e lasciato libero quindi utilizziamo quanto già definito per il path                    
        /downtime/status                                                                                
      tags:                                                                                             # ONLY EXTERNAL  
        - Status                                                                                         # ONLY EXTERNAL
      operationId: status                                                                               # ONLY EXTERNAL
      responses:                                                                                        # ONLY EXTERNAL  
        '200':                                                                                          # ONLY EXTERNAL
          description: Ok                                                                               # ONLY EXTERNAL
          content:                                                                                      # ONLY EXTERNAL
            application/problem+json:                                                                   # ONLY EXTERNAL
              schema:                                                                                   # ONLY EXTERNAL
                $ref: "./schemas-pn-components-v1.yaml#/components/schemas/PnStatusResponse"            # ONLY EXTERNAL
              example:                                                                                  # ONLY EXTERNAL
                functionalities: [NOTIFICATION_CREATE,NOTIFICATION_VISUALIZATION,NOTIFICATION_WORKFLOW] # ONLY EXTERNAL
                openIncidents: []                                                                       # ONLY EXTERNAL
        '500':                                                                                          # ONLY EXTERNAL  
          description: Internal Server Error                                                            # ONLY EXTERNAL
          content:                                                                                      # ONLY EXTERNAL
            application/problem+json:                                                                   # ONLY EXTERNAL
              schema:                                                                                   # ONLY EXTERNAL
                $ref: "./schemas-pn-components-v1.yaml#/components/schemas/PnStatusResponse"            # ONLY EXTERNAL
              example:                                                                                  # ONLY EXTERNAL
                functionalities: [NOTIFICATION_CREATE,NOTIFICATION_VISUALIZATION,NOTIFICATION_WORKFLOW] # ONLY EXTERNAL
                openIncidents:                                                                          # ONLY EXTERNAL
                  - functionality: NOTIFICATION_CREATE                                                  # ONLY EXTERNAL
                    status: KO                                                                          # ONLY EXTERNAL
                    startDate: "2019-08-24T14:15:22Z"                                                   # ONLY EXTERNAL
        '400':                                                                                          # ONLY EXTERNAL
          description: Bad request                                                                      # ONLY EXTERNAL
          content:                                                                                      # ONLY EXTERNAL    
            application/problem+json:                                                                   # ONLY EXTERNAL
              schema:                                                                                   # ONLY EXTERNAL
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'                         # ONLY EXTERNAL
  ###########################################################################################
  ###                                 INTEROP PROBING                                     ###
  ###########################################################################################
  "/interop/probing":
    get:
      summary: Provides the status of the E-Service
      description: >-
        Metodo che restituisce lo stato dell'E-service al probing di interoperabilita
      tags:
        - interopProbing
      operationId: getEserviceStatus
      responses:
        '200':
          description: The e-service is up and running
        '500':
          description: A managed error has occured during the request elaboration
          content:
            application/problem+json:
              schema:
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'

  ###########################################################################################
  ###                                 API PUBBLICHE                                       ###
  ###########################################################################################

  "/downtime/v1/resolved":
    get:
      summary: Recupero disservizi
      description: >-
        Metodo che restituisce l'elenco dei disservizi risolti per l'anno e il mese selezionato.
      tags:
        - Downtime
      operationId: getResolved
      x-api-cachekey-parameters:
        - "method.request.querystring.year"
        - "method.request.querystring.month"
      parameters:
        - $ref: '#/components/parameters/queryYear'
        - $ref: '#/components/parameters/queryMonth'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "./schemas-pn-components-v1.yaml#/components/schemas/PnDowntimeHistoryResponse"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'

  "/downtime/v1/status":                                                                                  # NO EXTERNAL
    get:                                                                                                  # NO EXTERNAL
      summary: Stato attuale                                                                              # NO EXTERNAL
      description: >-                                                                                     # NO EXTERNAL
        Metodo che restituisce l'elenco delle funzionalità di Piattaforma Notifiche e l'elenco            # NO EXTERNAL
        dei disservizi presenti al momento dell'invocazione.                                              # NO EXTERNAL
      tags:                                                                                               # NO EXTERNAL
        - Downtime                                                                                        # NO EXTERNAL
      operationId: currentStatus                                                                          # NO EXTERNAL
      x-api-permissions:                                                                                  # NO EXTERNAL
        - 'api-key-read'                                                                                  # NO EXTERNAL
        - 'log-downtime-read'                                                                             # NO EXTERNAL
      responses:                                                                                          # NO EXTERNAL
        '200':                                                                                            # NO EXTERNAL
          description: OK                                                                                 # NO EXTERNAL
          content:                                                                                        # NO EXTERNAL
            application/json:                                                                             # NO EXTERNAL
              schema:                                                                                     # NO EXTERNAL
                  $ref: "./schemas-pn-components-v1.yaml#/components/schemas/PnStatusResponse"            # NO EXTERNAL
              example:                                                                                    # NO EXTERNAL
                - functionalities: [NOTIFICATION_CREATE,NOTIFICATION_VISUALIZATION,NOTIFICATION_WORKFLOW] # NO EXTERNAL
                  openIncidents:                                                                          # NO EXTERNAL
                    - functionality: NOTIFICATION_CREATE                                                  # NO EXTERNAL
                      status: KO                                                                          # NO EXTERNAL
                      startDate: "2019-08-24T14:15:22Z"                                                   # NO EXTERNAL
        '400':                                                                                            # NO EXTERNAL
          description: Bad request                                                                        # NO EXTERNAL
          content:                                                                                        # NO EXTERNAL
            application/problem+json:                                                                     # NO EXTERNAL
              schema:                                                                                     # NO EXTERNAL
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'                           # NO EXTERNAL

  "/downtime/v1/history":                                                                                 # NO EXTERNAL
    get:                                                                                                  # NO EXTERNAL
      summary: Ricerca storico disservizi                                                                 # NO EXTERNAL
      description: >-                                                                                     # NO EXTERNAL
        Elenco dei disservizi di Piattafoma Notifiche riscontrate nel lasso di tempo specificato          # NO EXTERNAL
        dai parametri _fromTime_ e _toTime_. <br/>                                                        # NO EXTERNAL
        Il parametro _functionality_ è opzionale e ripetibile; permette di filtrare i disservizi          # NO EXTERNAL
        estratti limitandoli a quelli impattanti le funzionalità elencate.                                # NO EXTERNAL
      tags:                                                                                               # NO EXTERNAL
        - Downtime                                                                                        # NO EXTERNAL
      operationId: statusHistory                                                                          # NO EXTERNAL
      x-api-permissions:                                                                                  # NO EXTERNAL
        - 'api-key-read'                                                                                  # NO EXTERNAL
        - 'log-downtime-read'                                                                             # NO EXTERNAL
      parameters:                                                                                         # NO EXTERNAL
        - $ref: '#/components/parameters/queryFromTime'                                                   # NO EXTERNAL
        - $ref: '#/components/parameters/queryToTime'                                                     # NO EXTERNAL
        - $ref: '#/components/parameters/queryFunctionality'                                              # NO EXTERNAL
        - $ref: '#/components/parameters/queryPage'                                                       # NO EXTERNAL
        - $ref: '#/components/parameters/querySize'                                                       # NO EXTERNAL
      responses:                                                                                          # NO EXTERNAL
        '200':                                                                                            # NO EXTERNAL
          description: OK                                                                                 # NO EXTERNAL
          content:                                                                                        # NO EXTERNAL
            application/json:                                                                             # NO EXTERNAL
              schema:                                                                                     # NO EXTERNAL
                  $ref: "./schemas-pn-components-v1.yaml#/components/schemas/PnDowntimeHistoryResponse"   # NO EXTERNAL
        '400':                                                                                            # NO EXTERNAL
          description: Bad request                                                                        # NO EXTERNAL
          content:                                                                                        # NO EXTERNAL
            application/problem+json:                                                                     # NO EXTERNAL
              schema:                                                                                     # NO EXTERNAL
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'                           # NO EXTERNAL


    ###########################################################################################
    ###                          DOWNLOAD ATTI OPPONIBILI A TERZI                           ###
    ###########################################################################################

  "/downtime/v1/legal-facts/{legalFactId}":
    get:
      summary: Ottieni atto opponibile a terzi
      description: >-
        Fornisce le informazioni per scaricare un atto opponibile a terzi o, se tale atto va 
        recuperato dagli archivi, fornisce una stima per eccesso di quanti secondi bisogna 
        attendere.
      tags:
        - Downtime
      operationId: getLegalFact
      x-api-cachekey-parameters:
        - "method.request.path.legalFactId"
      x-api-permissions:                                                                                
        - 'api-key-read'                                                                                
        - 'log-downtime-read'
      parameters:
        - name: legalFactId
          in: path
          required: true
          description: Identificativo dell'atto opponibile a terzi che si vuole scaricare
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:  
            application/json:
              schema:
                $ref: "./schemas-pn-components-v1.yaml#/components/schemas/LegalFactDownloadMetadataResponse"
              examples:
                hot:
                  summary: File disponibile
                  value:
                    filename: downtime_NEW_NOTIFICATION_20221204.pdf
                    contentLength: 32423
                    url: "https://s3............com/presigned-url"
                cold:
                  summary: File archiviato
                  value:
                    filename: downtime_NEW_NOTIFICATION_20221204.pdf
                    contentLength: 32423
                    retryAfter: 7200    
        '204':
          description: Not signed
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'

    ###########################################################################################
    ###                                MODIFICA DELLO STATO                                 ###
    ###########################################################################################
  "/downtime-internal/v1/events":                                                                         # NO EXTERNAL
    post:                                                                                                 # NO EXTERNAL
      summary: Modifica stato di servizio                                                                 # NO EXTERNAL
      description: >-                                                                                     # NO EXTERNAL
        Permette di aggiungere un evento di inizio o fine di un disservizio, può essere                   # NO EXTERNAL
        richiamato sia dagli allarmi automatici che da GUI. Tali casi vanno discriminati                  # NO EXTERNAL
        con l'attributo di richiesta _sourceType_.                                                        # NO EXTERNAL
      tags:                                                                                               # NO EXTERNAL
        - DowntimeInternal                                                                                # NO EXTERNAL
      operationId: addStatusChangeEvent                                                                   # NO EXTERNAL
      x-api-permissions:                                                                                  # NO EXTERNAL
        - 'api-key-write'                                                                                 # NO EXTERNAL
        - 'log-downtime-write'                                                                            # NO EXTERNAL
      parameters:                                                                                         # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/uidAuthFleet'                                    # NO EXTERNAL
      requestBody:                                                                                        # NO EXTERNAL
        content:                                                                                          # NO EXTERNAL
          application/json:                                                                               # NO EXTERNAL
            schema:                                                                                       # NO EXTERNAL
              type: array                                                                                 # NO EXTERNAL
              items:                                                                                      # NO EXTERNAL
                $ref: "./schemas-pn-components-v1.yaml#/components/schemas/PnStatusUpdateEvent"           # NO EXTERNAL
        required: true                                                                                    # NO EXTERNAL
      responses:                                                                                          # NO EXTERNAL
        '204':                                                                                            # NO EXTERNAL
          description: OK                                                                                 # NO EXTERNAL
        '400':                                                                                            # NO EXTERNAL
          description: Bad request                                                                        # NO EXTERNAL
          content:                                                                                        # NO EXTERNAL
            application/problem+json:                                                                     # NO EXTERNAL
              schema:                                                                                     # NO EXTERNAL
                $ref: './schemas-pn-errors-v1.yaml#/components/schemas/Problem'                           # NO EXTERNAL

components:
  parameters:
    
    ###########################################################################################
    ###                          PARAMETRI DI RICERCA DEI DOWNTIME                          ###
    ###########################################################################################

    queryYear:
      name: year
      in: query
      required: false
      description: anno di ricerca dei disservizi risolti. Se non presente si intende l'anno corrente.
      schema:
        type: integer
        minimum: 2023
      example: 2024

    queryMonth:
      name: month
      in: query
      required: false
      description: mese di ricerca dei disservizi risolti. Se non presente si intende il mese corrente.
      schema:
        type: integer
        minimum: 1
        maximum: 12
      example: 3

    
    queryFromTime:
      name: fromTime
      in: query
      required: true
      description: data/ora di inizio dell'intervallo entro cui eseguire la ricerca
      schema:
        type: string
        format: date-time
    
    queryToTime:
      name: toTime
      in: query
      required: false
      description: data/ora di fine dell'intervallo entro cui eseguire la ricerca
      schema:
        type: string
        format: date-time
    
    queryFunctionality:
      name: functionality
      in: query
      required: false
      description: funzionalità di Piattaforma Notifiche di cui elencare i disservizi
      schema:
        type: array
        items:
          $ref: './schemas-pn-components-v1.yaml#/components/schemas/PnFunctionality'
    
    queryPage:
      name: page
      in: query
      required: false
      description: Pagina di risultati a cui il client è interessato
      schema:
        type: string
        
    querySize:
      name: size
      in: query
      required: false
      description: Size della pagina di risultati a cui il client è interessato
      schema:
        type: string
#  securitySchemes:           # ONLY EXTERNAL                           
#    bearerAuth:              # ONLY EXTERNAL                           
#      type: http             # ONLY EXTERNAL                           
#      scheme: Bearer         # ONLY EXTERNAL                         
#      description: JWT       # ONLY EXTERNAL                         
#security:                    # ONLY EXTERNAL                         
# - bearerAuth: []            # ONLY EXTERNAL                         
    